+++ resources/lang/en/app.php
@@ -96,4 +96,6 @@ return [
     'debug' => 'Debug',
     'view' => 'View',
     'user_witzh_access_to_sensitive_data' => 'Users with access to sensitive data',
+    'privacy_policy' => 'Privacy Policy',
+    'privacy_policy_agreement_link' => 'By using this service, you accept our <a href=":url" target="_blank">privacy policy</a>.',
  ];