+++ resources/views/auth/privacy.blade.php
@@ -1,0 +1,7 @@
+@extends('layouts.login')
+
+@section('title', __('app.privacy_policy'))
+
+@section('content')
+    {!! $content !!}
+@endsection